chore: move routes to respective `public` and `private` paths
