+++ .gitignore
@@ -1,5 +1,8 @@
 # See http://help.github.com/ignore-files/ for more about ignoring files.
 
+# environment
+.env
+
 # compiled output
 dist
 tmp
