+++ apps/assessment/middleware.ts
@@ -3,28 +3,26 @@ import type { NextRequest } from 'next/server';
 import { NextResponse } from 'next/server';
 
 export async function middleware(request: NextRequest) {
-  console.log('Middleware called for path:', request.nextUrl.pathname);
-
   const token = await getToken({ req: request });
-
-  // Allow access to authentication routes
-  if (request.nextUrl.pathname.startsWith('/api/auth')) {
-    return NextResponse.next();
-  }
-
   // Check if the user is authenticated
   if (!token) {
-    console.log('No token found, redirecting to login');
     // Redirect to login page if not authenticated
     return NextResponse.redirect(new URL('/login', request.url));
   }
 
-  console.log('Token found, allowing request');
-  // Allow the request to continue
   return NextResponse.next();
 }
 
-// Update the matcher to exclude authentication routes
 export const config = {
-  matcher: ['/dashboard/:path*', '/api/:path*'],
+  matcher: [
+    /*
+    * Match all request paths except for the ones starting with:
+    * - api/auth (API routes)
+    * - login (login page)
+    * - _next/static (static files)
+    * - _next/image (image optimization files)
+    * - favicon.ico, sitemap.xml, robots.txt (metadata files)
+    */
+    '/((?!api/auth|login|_next/static|_next/image|favicon.ico|sitemap.xml|robots.txt).*)',
+  ],
 };
